// src/app/interfaces/department.ts
export interface IDepartmentDoctor {
  id: string;
  img: string | ArrayBuffer;
  name: string;
  age: number;
  gender: string;
  address: string;
  role: string;
}


export interface Mybookings {
  id: string;
  reservations_id: 0,
  slote_id: 0,
  date: string,
  cust_id: 0,
  slote_no: 0,
  car_type: string,
  my_slote: string,
  amount_received: 0
}